+++ tomcat/src/main/java/org/apache/catalina/container/Container.java
@@ -1,0 +1,9 @@
+package org.apache.catalina.container;
+
+import com.java.http.HttpRequest;
+import com.java.http.HttpResponse;
+
+public interface Container {
+
+    HttpResponse service(HttpRequest request);
+}
